Greatly simplified coliders
Colliders were causing a bunch of lag, so I made them simplified and now the game runs pretty smoothly except for a short pause when chunks are loaded in. I'll probably add in threading or enumerators or whatever those are soon so that it doesn't freeze when  chunks are being loaded, and instead the framerate just drops for a second or two.